growity-creative-analyzer/backend/requirements.txt
modal>=0.73.0
fastapi
python-multipart
supabase
numpy
Pillow
ffmpeg-python
